requirements/base.txt
Django==2.2.4
pytz==2019.2
sqlparse==0.3.0
psycopg2==2.8.3
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
celery==4.3.0
flower==0.9.3
django-extended-choices==1.3.3
boto3==1.9.239
